removed test from UI tests, added to new class to avoid synchronization issue
