+++ test/DynamoCoreWpfTests/PackageManagerUITests.cs
@@ -9,9 +9,6 @@ using Dynamo.PackageManager;
 using Dynamo.PackageManager.UI;
 
 using NUnit.Framework;
-using System.Reflection;
-using System;
-using System.IO;
 
 
 namespace DynamoCoreWpfTests
@@ -94,38 +91,6 @@ namespace DynamoCoreWpfTests
             AssertWindowClosedWithDynamoView<PublishPackageView>();
 
         }
-
-        [Test]
-        public void AddingPackagesRaisesCanExecuteChangeOnDelegateCommand()
-        {
-            //set synchronization context since the package searchview from previous UI tests is still
-            //running and throws exception when attempting to modify searchresults
-            SynchronizationContext.SetSynchronizationContext(SynchronizationContext.Current);
-            var vm = new PublishPackageViewModel(ViewModel);
-            ViewModel.OnRequestPackagePublishDialog(vm);
-
-            //find a customnode to add to the package
-            string packagedirectory = Path.Combine(GetTestDirectory(ExecutingDirectory), "pkgs");
-            var packages = Directory.EnumerateDirectories(packagedirectory);
-            var first = Path.GetFullPath(packages.First());
-            string dyfpath = Path.Combine(first, "dyf");
-            var customnodes = Directory.GetFiles(dyfpath);
-            var firstnode = customnodes.First();
-                      
-            var canExecuteChangedFired = 0;
-            vm.SubmitCommand.CanExecuteChanged += ((o, e) => { canExecuteChangedFired ++; });
-            //now add a customnode to the package
-            vm.AddFile(firstnode);
-            
-            //assert that canExecute changed was fired one time 
-            Assert.AreEqual(canExecuteChangedFired,1);
-
-        }
-
-       
-
-
-
         #endregion
 
         #region InstalledPackagesView
